Rename `initialized` to `run_initialized` and move process message
