+++ test/addon_test.rb
@@ -35,7 +35,7 @@ module RubyLsp
 
     def test_registering_an_addon_invokes_activate_on_initialized
       server = RubyLsp::Server.new
-      server.initialized
+      server.run_initialized
 
       addon_instance = T.must(Addon.addons.find { |addon| addon.is_a?(@addon) })
       assert_predicate(addon_instance, :activated)
